refactor(compiler-plugin): extract HintEntry computation into PreviewLabHintEntries
Address PR #164 review threads from TBSten:

* Pull `HintEntry` data class plus the cache + `computeHintEntries()` logic
  out of `PreviewLabHintFirGenerator` and into a dedicated
  `PreviewLabHintEntries.kt`. The generator now only owns the
  `FirDeclarationGenerationExtension` overrides; the rule that turns a
  `FirSession` into hint entries lives next to its data class.

* Drop the explicit `Companion` name on the private companion object —
  `Companion` is the default and adding it added no information.

* Normalize the `**Hint function — Before**` heading in
  `PreviewLabHintIrBodyFiller`'s KDoc into a "**Hint function**" section
  with separate "Before" / "After" lines so the Before/After block reads
  like the rest of the compiler-plugin KDoc style.

The `hasPackage(FqName)` override stays as-is: it overrides the
`FirDeclarationGenerationExtension` API contract so renaming it is not an
option, and the override already does exactly what the parent contract
asks for ("does this extension contribute anything to this package?").
